feat: 1.添加拦截器 2.用户资料修改接口 3.调整swagger

+++ src/main/java/com/example/red/book/model/vo/UserVO.java
@@ -12,9 +12,9 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 public class UserVO {
 
-    @NotNull(message = "id不能为空" )
+    @NotNull(message = "id不能为空")
     @ApiModelProperty("用户id")
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty("用户名")
     private String username;
